remove Page component
