--- app/src/components/Page/style.scss
@@ -1,73 +1,0 @@
-@import "~style/vars.scss";
-
-.page {
-  margin: 0 auto;
-  padding: 20px 45px;
-  //max-width: 960px;
-  background-color: whitesmoke;
-  color: black;
-  display: flex;
-
-  .seperator {
-    width: 1px;
-    margin-top: 100px;
-    background: rgb(100, 100, 100);
-  }
-
-  .section {
-    padding: 28px 12px 0 12px;
-    position: relative;
-  }
-
-  .market-section {
-    flex-basis: 70%;
-  }
-
-  .position-section {
-    flex-basis: 30%;
-
-    .heading {
-      margin-top: 118px;
-    }
-  }
-
-}
-
-/*
-.invest {
-  width: 100%;
-  font-size: 24px;
-}
-
-.predictions {
-  position: relative;
-}
-
-.lockedOverlay {
-  display: none;
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  text-align: center;
-  z-index: 100;
-  vertical-align: middle;
-  background-color: whitesmoke;
-  opacity: 0.8;
-
-  &.lockedOverlayEnabled {
-    display: inline-block;
-    font-weight: 800;
-
-
-    &::before {
-      content: '';
-      display: inline-block;
-      vertical-align: middle;
-      height: 100%;
-    }
-  }
-}*/
